Remove some duplicated functions
