+++ src/wazuh_testing/constants/paths/binaries.py
@@ -18,3 +18,4 @@ AGENT_AUTH_PATH = os.path.join(BIN_PATH, 'agent-auth')
 ACTIVE_RESPONSE_BIN_PATH = os.path.join(WAZUH_PATH, 'active-response', 'bin')
 ACTIVE_RESPONSE_FIREWALL_DROP = os.path.join(ACTIVE_RESPONSE_BIN_PATH, 'firewall-drop')
 MANAGE_AGENTS_BINARY = os.path.join(BIN_PATH, 'manage_agents')
+AGENT_GROUPS_BINARY = os.path.join(BIN_PATH, 'agent_groups')
