+++ src/wazuh_testing/constants/ports.py
@@ -1,1 +1,5 @@
+# Copyright (C) 2015-2023, Wazuh Inc.
+# Created by Wazuh, Inc. <info@wazuh.com>.
+# This program is free software; you can redistribute it and/or modify it under the terms of GPLv2
+
 DEFAULT_SSL_REMOTE_ENROLLMENT_PORT = 1515
