+++ src/wazuh_testing/utils/db_queries/global_db.py
@@ -80,10 +80,3 @@ def clean_agents_from_db():
         database.query_wdb(command)
     except Exception:
         raise Exception('Unable to clean agents')
-
-
-def list_agents_ids():
-    wazuhdb_result = database.query_wdb('global get-all-agents last_id -1')
-    list_agents = [agent['id'] for agent in wazuhdb_result if not (0 == agent.get('id'))]
-
-    return list_agents
