+++ src/wazuh_testing/utils/manage_agents.py
@@ -11,7 +11,7 @@ from typing import List
 from wazuh_testing.constants.api import AGENTS_ROUTE
 from wazuh_testing.constants.paths.binaries import MANAGE_AGENTS_BINARY
 from wazuh_testing.modules.api.utils import login, get_base_url
-from wazuh_testing.utils.db_queries.global_db import delete_agent, list_agents_ids
+from wazuh_testing.utils.db_queries.global_db import delete_agent
 
 
 def remove_agents(agents_id: List, remove_type: str = 'wazuhdb') -> None:
@@ -47,7 +47,3 @@ def remove_agents(agents_id: List, remove_type: str = 'wazuhdb') -> None:
             response_data = response.json()
             if response.status_code != 200:
                 raise RuntimeError(f"Error deleting an agent: {response_data}")
-
-
-def remove_all_agents(remove_type):
-    remove_agents(list_agents_ids(), remove_type)
